Rework fix for SCC algorithm
- revert most of the changes of 8231a671 because they introduced a change in
  the API and unnecessarily traversed nodes multiple times. However the commit
  discovered one core issue, which was during cycle detection the update of
  the SCC id with the start number (wrong) rather than the scc id (right).

  This fix, along with correct handling of cached values is now implemented
  ontop of the old code-base.
- Merged the "AdvancedSCCTest" with the regular one.
- Refactored internal variable names to be more precise and added JavaDoc/
  comments on what (some of) the individual variables/methods actually do.
